+++ setup.py
@@ -5,7 +5,7 @@ setup(
     entry_points={
         'console_scripts': [
             'generate_data = training_scripts.cli:generate_data',
-            'generate_balanced_dataset = training_scripts.cli:generate_balanced_dataset',
+            'select_balanced_dataset = training_scripts.cli:select_balanced_dataset',
         ],
     },
     install_requires=[
